Thrown summon sphere cards explode on contact.
This is just for cartomancers, but I thought it would make them really fun to play. They basically get 'bomb' cards whenever they get a flaming sphere or any other sphere card. If the card misses it still summons the monster, but if it hits, it explodes for the elemental type instead of summoning a monster.

Soon we'll import stinking spheres from Fourk for another explosive card.