Monsters can read scrolls of water.
This is probably gonna be nasty - but at least this only a defensive measure for them. The scroll works exactly for monsters as it does for the player. We'll see if this is as bad - or worse - than the Make Pool spell from SLASHEM and reevaluate, but it should provide for some interesting situations.