Fix: eel hiding out of water (fmon) on quest filler levels.
This was happening because I added rivers to some quest levels - unfortunately including it in all quest levels that counted as fill levels. This was more than I counted on. The fix is simple though - just omit the entrance and nemesis levels.

After re-checking some of the quest levels, I really like the rivers change, it definitely adds variety and challenge to some quests, especially the ranger quest where we mix a lot of rivers and trees together.