Spell of charm monster nerfs and updates.
Adapted from YANI #3991 (https://nethack-yanis.github.io/yanis/3991.html).

Charm monster is less powerful but back to being a level 3 spell. It is now directional and only affects the first monster it hits. For unskilled and basic users, it will only pacify monsters. It will tame monsters when you are skilled or better in enchantment spell school.

The recent changes regarding pets and charisma definitely nerfs this spell even more, that is the primary reason I lowered it back to 3. However, being able to pacify monsters easily is still powerful.