Flanking bonus overhaul.
A couple things about the way flanking was calculated was still bugging me. If any monster in flanking position got the same bonus formula, it didn't really matter much if the monster was a natural flanker or not. I have changed that so that outflankers get a much higher bonus for doing their job. Also, as a bonus for the player, certain roles that are more fighting oriented also count as flankers (knights, valkyries, etc) and get the higher bonus.

What tipped me off to this in the first place was noticing that some low level monsters got lower than a 4AC bonus and minotaurs only got ~10AC for flanking. Now, minotaurs can get a very scary +18-+20AC bonus, something that can make quite a difference in the endgame.

The minimum flanking bonus is 4AC and the max is currently 20AC. It's possible I'll adjust these after playtesting.