Preventative fix for infinite swap weapon loop.
A strange infinite lock has cropped up in the fuzzer a few times. The player will have a shield on and attempt to swap in a two-handed weapon, and it gets stuck in trying to wield the weapon. I'm very certain it's being caused by making weapon swapping take 0 turns but not quite sure. I'm going to add a couple commands to clear the repeat queue - however I was not able to reproduce the infinite loop so it remains to be seen if this fix is successful.