Fix: trapped without a trap (fmon) when polyd into a green slime.
We could be hidden (on the ceiling) as a slime, next to a monster that is trapped in a web. The monster attacks us, then we drop from the ceiling to surprise attack it and in the process this causes the monster to be displaced by us. The mtrapped flag is not removed causing the error.

My solution was to block displacing if the monster is trapped, since in theory it probably shouldn't be able to move.