Mind flayers won't usually attempt to eat the brains of petrifying monsters.
This arose from an attempt to fix a mind flayer continuing to attack a basilisk after being stoned, but it led me to question the concept of supposedly smart flayers even trying to attack in the first place. This commit makes flayers consious of petrifying targets and they won't attempt tentacle attacks unless impaired.

This also applies to the player! So it's a nice quality of life feature (unless you are confused, stunned, hallucinating, or rabid).

The next step is to prevent intelligent monsters from attacking stoners bare handed, because currently even though a flayer won't eat the brains of a cockatrice, it will punch it bare-handed :P.