Tweak odds for #offer odds (from EvilHack).
I noticed this on the EvilHack ideapad, and it was something I already was thinking about - mostly because in this variant we have no other lesser gifts. It's all artifacts or nothing. But the odds looked like a large improvement from previous, so we're rollin with it. Thanks for doing the math for me K2!