Repair spell (adapted from repair armor spell in EvilHack).
Couple of tweaks, for one, this spell can be used on any item, not just armor. With so much item damage occurring (especially corrosion from acid), it would be invaluable to have a way to deal with it.

Also, at skilled, this will remove all damage instead of 1 level of erosion at basic.