Cartomancer spellcasting overhaul.
This is something I mulled over for a long while. Cartomancers have a brand new mechanic that revolves around cards. However, if we allow them to become really good at spellcasting, do they not just become pseudo-wizards? I think we need to preserve the core mechanic that cartomancers are designed around and significantly weaken their spellcasting powers - this makes them continue to use "the cards that they are dealt" through the mid-and-endgame.

This commit is heavily influenced by YANI #3866 by spicycat (https://nethack-yanis.github.io/yanis/3866.html). It took me a while to come around to this idea, but the more I thought about it, the more it made sense. Downgrading spellbooks into consumable wand-like objects does a couple of things.

1) It lets cartomancers cast from spellbooks much more freely than other roles can. Like zapping cards, carto's start out knowing the identities of all rulebooks since usually the title and purpose of a rulebook is quite clear on the cover. Also, cartomancers don't have to actually spend time reading and learning the spell - they can just cast it with a brief glance at the rules. And finally, cartomancers will always cast spells at expert level since spellbooks are consumable and they also have access to the fine print.
2) It limits the power of all spells since cartomancers can't permanently learn any spells. For all intents and purposes, spellbooks are pretty-much wands but with the extended variety of spell-types.
3) Cartomancers will have to rely on their summon, zap, and razor-cards to survive.
4) Cartomancers will (hopefully) play out much more differently than wizards late-game.