Adjust prices of cheaper scrolls..
This is a minor change designed to fuzz price-id a little bit. I do have a long term vision to make price identification obsolete, but that's for a later time.

blank paper	was 60		now 50
enchant armor 	was 80		now 200
enchant weapon 	was 60		now 200
remove curse 	was 80		now 300

I won't altar the ink costs because magic markers were already made much more rare. This change does make the player have to gamble a bit more on the most dangerous class of scrolls, IMO, the 200 scrolls which include amnesia.