Archeologists get built-in price identification.
Adapted from YANI #1539 by FIQ (https://nethack-yanis.github.io/yanis/1539.html).

Archeologists are a pretty weak starting character, so I think they aren't as popular as they should be. But with this change, you get to have the benefits of price identification without even needing a shop! This reduces tedium and makes the role have an outstanding and unique feature.

Currently the exact base price is shown in parentheses for all unknown items except gems and stones. I left out those because the arc starts with a touchstone.

Potential extensions:
* Remove the price if the player has type-named the object. I'm not sure if this would help or not, but it might prove useful.