Falling rock traps can dump multiple rocks.
Adapted from YANI #3977 by aosdict and ais523 (https://nethack-yanis.github.io/yanis/3977.html).

As the level difficulty increases, the chance of multiple falling rocks and boulders also increases. The quantity is determined by level_difficulty / 4. Boulders will be half the quantity. I decided not to apply the same effect for when monsters walk into falling rock traps since the dungeon is supposed to be more dangerous for the player.