Cartomancers don't get a starting pet and cannot tame monsters.
One last push to make the carties more reliant on their cards. This may be kind of controversial, since pets are an important part of other roles, but the cartomancer DOES have their own version of pets - and quite a diverse and power variation of it. The cartomancer also has an overriding theme of comsumerism and using lots of one-off items instead of having reliable long term items, so I'm going to experiment with this.

The problems I've been noticing is that a carto can let their pet kill everything in the beginning while collecting all their cards and build up a huge stash and be quite spoiled by minetown. This also creates inventory management problems (too soon in my opinion) and can clog the game up quite quickly. If a player doesn't start with a pet and can get a powerful one later, than summon cards still lose most of their value later and start becoming junk.