Cartomancers can cast card combos.
This is adapted from the Card Combo tech from the old SpliceHack, which was ported to HackEM but never realized.

At level 5, cartomancers will be able to use the Z command to execute card combos.
At first they will only be able to cast 3 cards at a time, but as they gain levels they will be able to cast more.

Similar to techs, there is a timeout of 1000-1500 turns. This is the only feature that will come close to resembling techs.