Water dipping symmetry (from SLASH'EM).
There is an odd assymetry in NetHack where if you dip a potion into water it dilutes, but if you dip water into something (other than acid), it results in random alchemy.

This is one of the strange cases where SLASH'EM got it right, but Vanilla NetHack never decided to port the change, probably deciding to go more with the alchemy nerfs they introduced over 3.6.