Do not autopickup when hallucinating.
This is an attempt at a fix for a heap-buffer-overflow when an item is auto-picked up. I believe it is connected to my recent change making hallucination affect all items. This also just makes hallucination much less dangerous in general for picking up things like heavy items or cockatrice corpses.