Allow pounding when blind.
This is a response to issue #611 on the NetHack issues. There might be some cases where this really stretches the line of game reality, but it seems like a good way to make polearms a little better. If monsters are in the way, we assume the hero is attacking around them.