Elbereth fades instantly for vampires and orcs.
I thought about the mechanic in EvilHack where infidels are considered cowards if they fight on Elbereth, but I decided it was just easier for the magic of Elbereth to just not be available. The coward thing also requires a coward flag which I didn't feel like messing with. This makes vampires a little more vulnerable, but also more thematic.