Ported shoggoths (from SLASH'EM).
These are a nasty threat in the late-game unless you have a source of acid resistance (which is quite difficult to find). I will not be importing the giant shoggoths, they are a bit too powerful for the game - as a foe and as a polyform. A design decision was not making these drop globs like the other P do. The globs require a few more moving parts, but I'm not sure these would be globby, so they'll simply drop a corpse.

One key modification: they are now mindless and cannot be detected with telepathy. This goes with the other changes with piercers, trappers, and lurkers above.

They also do not have the traitor ability from SLASH'EM.