Yet one last tweak to card drop frequency.
Okay - I think I have the sweet spot now. I even made an Excel spreadsheet to show the distribution of cards based on your level and the dungeon level you are on. Ideally the player gets a pretty generous distribution of cards at the beginning of the game up until reaching XP 6-8, when the cards start to thin out a bit. Hopefully this is the point at which the cartomancer is able to start fending for themselves and has a decent backlog of cards to use, or they've found a suitable long term weapon, or backup projectiles, or maybe they are very creative and figured out some other way to survive.

The card drop frequency is more based on the players level than the dungeon level so they can't just teleport back to level 1 to farm cards - although, one could still do that with favorable results. I don't think that's too exploitative though and could be a viable strategy if you run out of cards in the late game.

Hopefully this is the last tweak, it will take some more playtesting though.