Cartomancer quest artifact overhaul.
The artifact #invoke from SpliceHack has not sat well with me for a long time - I thought we could do better and make it actually work with the core mechanic of the cartomancer instead of just spewing out a bunch of random monsters.

Now, the Holographic Void Lily can be invoked to become lit for a short period. During this period, the cartomancer will enjoy a much higher rate of card drops - the rate will actually be inverted from what they normally experience. So if the rate was 1 in 3 monsters drop cards, it is now 2 in 3. The standard duration of the invoke is 25-50 turns. This is my initial guess for a good time, but it has not been playtested. I figured that would be enough time to collect a few cards, or many if you are in a monster heavy location.

At first, I thought I would need to add another flag (like a cooldown flag) to enable this mechanic. Actually lighting up the card didn't occur to me until later, but it works perfectly and eliminates the need for another instance variable. The Lily does not count as artifact light, so I had to work in some extra handling for the lighting.

I'm pretty happy with this idea and I think it also makes sense for the artifact itself. Since it's based on the Black Lotus, a card that grants the player powerful resources, this artifact now functions in a similar fashion.