Unicorn horns cannot cure Rabid status - new cures in it's place.
I did some *very* minimal research to come up with some fun new cures for rabid, here we go:
* garlic
* dog liver (meaning you can eat any dog corpse)
* rubbing a non-cursed foulstone (ie: "madstone").

Two of these cures are consumable - the third is not quite, but when the foulstone is used this way, it autocurses. I didn't like that the unihorn could just cure rabid like everything else, so now we have some different extra ways to deal with it.