Mind flayer blasts do more damage and can stun the player.
Bumped the standard damage from 1d15 to 5d6 - effectively doubling it. Master mind flayers also deal an extra 3d5 damage.

For the icing on the cake, if the player takes more than 6 damage from a psychic blast, they are also stunned.

This might seem a little OP, but keep in mind it still costs 20 energy to use #monster. Perhaps a role with an artifact that regenerates energy could abuse it, but seems like the exception rather than the rule. This also makes the mind flayers more dangerous whether they are close to you or not. I don't recall ever really being threatened by the psychic blast, so hopefully this makes it a bit more interesting.