Double down on cartomancers weakness in melee combat.
While playtesting I came across a random corroded +9 war hammer. Despite the existing -10 to-hit penalty, I was still able to pick it up and consistently bash with it. For any other role this is, of course, acceptable. But for cartomancers it gives them an out from the card drops. I also found that as soon as I obtained gauntlets of power the game became much easier and my reliance on card drops dropped dramatically.

To carve the cartomancer's path back toward drops I took these steps (all of these penalties apply only to melee combat - NOT ranged attacks):
* Add a -10 to-hit penalty for wearing a shield.
* Add a -20 to-hit penalty for wearing armor.
* If wielding a weapon, an additional -10 to-hit penalty is assessed.
* Lowered the max bare-handed skill to basic.

This means that a player could eventually use just gauntlets of power or a really nice (and heavily enchanted) weapon, but they will need to work up to it. Other gauntlets might be better options, but GoP is still fantastic for the carrying capacity.