Cartomancer card drops can include random NOGEN monsters.
This is something that totally got past me, but there was an inconsistency in the starting summon cards that cartomancers can receive and the random drops. They each used totally different methods of generating the actual summon. I have updated the random drop formula to include some totally random monsters, including NOGEN monsters, to spice things up a bit. What this means is that before, we were not able to get random stuff like minotaurs or woodchucks unless we actually killed one and got the rather small chance of their same card drop. I also opened up the summons a bit in the last commit for del zethire and mplayer cartos, now that change will apply to random drops as well.

Something I sorely noticed missing from the random drops in general was demons, which can be very good allies. The previous formula would only allow them whilst in Gehennom, but now at least there is a small chance of seeing them drop in the main dungeon.

I'm still not sure on how to approach unique monster cards. I really like the idea, and it probably just requires jumping in, bugs and all...