Cartomancer quest start re-design.
Nothing too crazy - I envisioned a circle of trees around the card school. It turned out more of a tree wall, but it's a little more interesting to explore. The player will have to figure out how to get through it, the trees are choppable.  I also shrunk the school a little bit and added more S and N monsters.