Playing cards costs energy for the cartomancer.
I figured that since energy was being used for almost nothing, this would be a good way to give the cards a little leash and the energy bar will at least be a little more active.

Playing each summon card costs 10 energy. Right now this seems the most reasonable, it's equivalent to a level 2 spell. A good cartomancer needs to also figure their energy in when entering a fierce battle.

I'm still not sure about the zap cards - maybe they should cost the same.