Cartomancers start with 7 summon raven cards. This is sort of my homage to the idea of giving cartomancers a pet raven. Since it will never happen (cartomancers should and will never receive proper pets), they can have this ominous starting hand as a reminder of what can never be.
This also eliminates an obvious start-scumming strategy
of being able to retry for 7 random high level monsters in
your starting hand to ensure the ability to destroy the
minetown priest (or another worthy target).

And finally, this also frees up a few precious inventory
slots, which I noticed can be an annoying issue for early
cartomancers once they've received a large number of cards.