Decks of cards can be tipped and emptied for razor cards.
This felt like a natural extension of these items, and makes them more useful for the cartomancer if they need more cards.

Might also be neat for other roles that run into the deck of fate and don't want to risk using it.

Any deck can yield 22-52 razor cards. The BUC and enchantment are up to the RNG as usual.