Further focus cartomancer skills.
After playtesting a lot, I've noticed I tend to rely on a stack of daggers as my goto. This is okay, but its kind of boring and not how I intended the role to be played. In the previous commit I fixed the card drop frequency to be much more generous, so I think we can in turn remove the other permanent ranged options the carto has and still further push them into cardplaying.

Actions taken:
* Remove dagger, knife, sling, quarterstaff and boomerang skills entirely.
* Increase max skill of unicorn horn to expert
* Increase max skill of darts to skilled.

* Also increase the starting razor cards to +2 so they last longer.