Summon cards are much more common for cartomancers.
This is another item generation hack that I think will make the gameplay more interesting.

I tweaked the core makeobject probabilities a little in favor of scrolls, I doubled the probability of cards of create monster, and I made card of create monster usualy spawn keyed to a specific monster(19 of 20 to be specific).

This means that we'll be able to find more immediately usable cards lying around the dungeon, but also that more monsters will be using them against us.