Refactor the restrictions on spell-beings; fix some troublesome summons.
I basically unified the code, now that we have many places around the codebase where spell beings are generated. In particular I noticed that watchmen/captains could become angered as spell beings, so they are currently outlawed, but maybe in the future we'll let them be spell beings but handle them with more care.