Fix: Intrinsics from wielded artifact/crystal ball could remain after stolen.
I noticed this when playtesting the cartomancer with the crystal ball got stolen, but further testing revealed it also happened with Quick Blade, Tsurugi, Serenity, and Werebane. These are the artifacts we didn't have enough room for special artifact flags so had to handle them in a less elegant fashion - and here we are paying the price.

I was able to handle the crystal ball problem and I refashioned some unused SPFX tags so the artifacts are handled properly now.