You cannot fly over falling rock traps.
For vampires, this was yet another trap that they are immune to, so I'm disabling that so they can enjoy the rocks. But it also seems weird that you could fly "over" a falling rock trap, we'll make the assumption that there's an infrared trigger rather than a floor trigger.