Buff Load Brand.
I was doing some damage calculations with the artifacts and noticed it wasn't nearly as strong as its weight should command.

* Buffed from 1d14 vs small to 3d10.
* Buffed from (1d8 + 2d6) vs large to 3d16.
* Instead of +1d40, it's simply double damage.

For a +0 Load Brand, you can expect 42 damage on average. That's almost equivalent to a +13 Grayswandir in the perfect situation. A +13 Load Brand is a nightmare to behold and deals over double what Mjollnir would.

It's heavy, but now it's probably worth it.