Fix: Paranoid trap confirm when displaced onto a trap.
Occurs when a displacer beast is occupying a space where a trap is, but not trapped (ie: crawled out of a pit trap). If we are displaced onto the trap, we should not have a say in whether we move or not.

This doesn't seem to be an issue upstream, probably caused by some of the extra displacement code.