Remove reflection message when blind.
This caused a strange sequence when the player had full shock resistance:
"You appear to only be partially affected.  You aren't affected."

The lack of a message when blind should be fine - in the case of shock res it provides a message already and other intrinsics should do something similar.