Anti-magical items resist being enchanted (from SpliceHack).
I discovered this gem on one of the SpliceHack-rewrite pull requests. It softly nerfs already powerful items but also subtly decreases the player's AC. This includes the cloak of magic resistance, shield of anti-magic, and even artifacts like the Sceptre of Might. I can hear the cavemen groaning...

Unlike the Splice PR, this doesn't completely eliminate being able to enchant antimagic items. Instead the odds are just significantly reduced as the enchantment increases:
* +0 = 1 in 2 chance of success
* +1 = 1 in 3 chance of success
* +2 = 1 in 4 chance of success.

So, it's not impossible to enchant these items, you will just have to burn through a lot of enchant weapon scrolls to achieve it.