Silver dragon scales are actually coated in silver.
This is another neat idea I found on the SpliceHack PR's by RojjaCebolla. I didn't change the actual material of the scales as the PR intended, but merely handled it in is_silver checks so otherwise the scales stil function as normal.

I would followup with a gold-coated scales commit, but there doesn't seem to be much point since I can't think of any effect on gameplay it would have.