Initial port of Lethe Patch levels.
This was a project I experimented with in HackEM but ultimately decided not to implement. However, I had cleaned up the original Lethe patch and combined in some updates from dnethack. I think there will actually be a good place for the Lethe levels in NerfHack. Instead of branching them separately, I think the lethe levels will fit after Medusa's and before the castle, and the Gulf of N'Kai can be near the end of Gehennom.