Lethe levels are very suppressive of hero gains.
This is my twist on the lethe water (which most players seem to hate). Lethe water can ruin a run in one mistake, blanking your entire inventory, so that's not fun.

What this does instead, is something more subtle, but not as dramatic as mass cancelling. Instead these effects take place on lethe levels:
* You cannot gain experience.
* You don't gain skill points for using skills
* You cannot read spellbooks to learn spells.
* You cannot engrave Elbereth, you instead forget how and become confused.
* There are no extra death drops from monsters.