Port Cerberus from EvilHack.
This started out as just enabling Cerberus from Vanilla, but the EvilHack version is much stronger and more fun. I didn't port the travel restriction, so currently it is not mandatory to kill Cerb to go to the VotD. I will probably leave that out and let Cerberus simply be an end boss for the Lethe branch.