Port the Bridge of Khazad-Dum as a welcome room to hell.
I've had this idea bouncing around for a long time and I've envisioned this level as the perfect level just after the Valley of the Dead.

If you cannot fly or firewalk, you are forced to take the narrow path in between the lava. This could be quite treachorous on the ascension run when nasties are summoned. Even if you can go over the lava, firewalking is dangerous for items, and flying still leaves you vulnerable to being surrounded.

Just so it's not so on-the-nose, I removed Durin's Bane. The level is now fairly quiet and a nice reprieve after clearing out the valley.

I had a little extra fun and added the fumaroles from the plane of fire here. A big difference is that we can actually see most of them here. Players are usuallly so stressed on the planes that they probably don't see how they generate.