Even MORE cracked altars!
Now that I *cracked* the lua code for getting cracked altars in special levels, many of the guaranteed altars in quests now have a good chance to already be cracked. This is especially true if they are unattended or unaligned, implying there was recent conflict and the altar was damaged.