Fix a couple things with hiders popping out from adjacent squares.
It's most important to check the square we occupy (to see if there is a hider on top of us) - so that is now checked first. There was also a chance for our pet to surprise us from the ceiling on turn 1 - I think this was from the mundetected not being checked correctly, so now that should not happen.

If there is an adjacent hider, we now just move them over our square so they can be handled like we walked into them.