Nerf slow digestion.
Adapted from YANI #2023 by aosdict (https://nethack-yanis.github.io/yanis/2023.html).

Slow digestion is incredibly powerful - the ability to reduce your hunger rate by a factor of 20x enough to almost bypass food altogether.  It also enables much easier altar camping. It also gives vampires a (practically) free pass once they get a "oSD or white dragon scale mail.
This feels like something that really needs to be nerfed - so we're going to do it. Like the YANI says, we're implementing the opposite effect of a ring of hunger. I wanted to keep this as simple as possible, so I used a random coin flip each time slow digestion is checked. I could have moved things around a bit so that it was actually checked every other turn, but in the long run it should have the same end result.

My hope is that this is still a great help, but not so powerful it lets players muddle around for thousands of turns.