The Lethe levels can have up to 4 altars generate.
This will sort of be the hidden treasure of the lethe. This kind of uses what dnethack added for altars, but I was much more restrained.

No altars are guaranteed. There are 4 levels where altars can show up. Where there was an altar specified, I gave it a 50% chance to generate. If it does spawn, there is a further 75% chance it generates cracked. If the altar doesn't generate, a pile of rocks spawns in its place.

There are no temples - the caretakers fled long ago...