Polymorphing into a horned monster destroys flimsy or cloth helms.
One time while playtesting I noticed that there was a message for polymorphing into a horned monster with a helm (ex: Your horns pierce through the elven helm), but it didn't actually damage it. After looking into it I noticed a note from the devteam in polyself.c:breakarmor(). It only took 2 lines to complete the effect and it makes sense, so it shall be.