Monster spellcasters can cast vs other monsters (from EvilHack).
In this update, I adapted monster spellcasting in mcastu.c so that monsters can effectively target other monsters. With the cartomancer role, we are exposed to many more interesting pet types - many of which are spellcasters. Not having these summoned pets be able to cast spells is a huge detriment to the role, but it also makes these types of pets more fun to play. Even if they aren't tame, conflict can make them more fun in a party.

I took the opportunity to do a bunch of variable renaming during this overhaul. The variable names in mcastu.c leave much to be desired - with mtmp usually being the caster, I instead used caster for the casting monster and mdef for the defender. This makes reading the code much easier.

I also decided to fold the changes in from EvilHack to the existing functions instead of copying the EvilHack implementation. There is a lot of code duplication and I was able to merge the separate player and monster handling in cast_wizard_spell and cast_cleric_spell fairly easily. I still kept castmm and mspell_would_be_useless separate for clarity though.

I left out the 'player can cast monster spells when polymorphed' ability. Three reasons: this is more complicated, it's a niche use-case, and it's probably overpowered. It lets roles that can't cast spells (ie: caveman) easily start casting spells and gives cartomancers a different avenue to avoid solely playing cards.

For spells that felt too strange or powerful for monster vs monster interaction, I either skipped them or turned it into a different effect. In general, I don't think we want monsters summoned more monsters when fighting eachother. It brings up some weird questions - like are the new summons tame? And can they be abused for altar camping?