Reveal mimics hit by sleep rays.
This fixes an inconsistency that the sleep ray has with other rays hitting mimics.

Most rays will uncover a mimic when hitting a mimic but not the sleep ray, it will put the mimic to sleep without unhiding it. This resulted in a message that described the mimic being hit without actually seeing it.