Stop using mksobj for pseudo wands.
This is setup for the next drastic change to cartomancers, but it also might be a good refactor. In some places, a temporary object is created when a wand effect is needed (Mjollnir's zap lightning invoke, or all zappable cards). I have been using mkobj to create a temporary object, but it seems there is a cleaner way using a plain obj struct.