Monsters can hide under grass and other dungeon furniture (from xNetHack).
This combines a few commits from xnh and updates it for 3.7 compatibility.

Commit 705adfd18 from xnh:

'Grass, headstones, altars, sinks, ladders, and thrones all provide
enough cover for a monster that can hide under objects, even if there
are no objects there. They behave exactly the same; for instance, a
concealable monster that steps onto a sink will vanish under it and
become undetected.

The AI for concealable monsters is tweaked so that in large areas of
concealing terrain such as grass fields, they will pick their spot to
move and *then* decide with their 90% chance not to move, *if* the spot
would be moving out of concealment. Previously this was tested before
they actually looked for new places to move, meaning that a monster in a
field of grass would be pointlessly sluggish.'

Updates in NerfHack:
* Monsters can also hide under forges, fountains, and toilets (eek!).
* Handled unhiding monsters when furniture is destroyed.