Fix: mhitm spellcasting damage wasn't handled properly.
I noticed I took some mysterious damage during playtesting while my winter wolf was attacking a minetown priest. This turns out to be due to the damage handling in the mage and clerical spellcasting functions. I reworked them slightly so that the monster damage is returned to be handled in castmm.