Increase carry capacity.
Previously, I nerfed one of the major aspects of the strength attribute, which is that it would grant significant to-hit bonuses. To compensate for that, I think allowing strength to have a greater effect on carrying capacity would be really beneficial. Couple reasons for this: the dungeon is longer and there is more item collecting and travel required.

I've thought about going down the path of implementing magic chests or thiefstones, but I don't like those. I think this method is more interesting (not to mention simpler), and rewards building strength.

Now, with the generous rings of carrying, it's possible to get up to 2500 carry cap with maxxed strength and two +5 rings of carrying.