Puddles (from EvilHack).
I adapted most of the implementation of puddles from EvilHack. This was quite a massive undertaking, but I think puddles are a great feature and a great nerf. Having access to unlimited water dilution is powerful, and puddles help to limit that resource. Permanent water pools will still appear early in some circumstances, like the mini-castle minetown and around some vaults, but otherwise the intention is to keep that resource to a minimum until the later levels.

Summary of puddle features:
* Puddles can be wished for
* Fountains gush out puddles instead of pools
* The scroll of flood usually creates puddles (it only creates a pool if cursed)
* Puddles can be used for dipping items into
* Puddles can rust metal shoes
* Tiny monsters can drown in puddles
* Rust traps leave a puddle when disarmed
* Whenever a rusting effect could occur, it can also dry up the puddle.
* Gremlins cloning in puddles also can dry up puddles.

Some changes and updates from EvilHack:
* Currently puddles don't damage long worms, this may be ported in the future.