KEEN vs MAX_KNOW.
KEEN is how many turns the player gets for a new spell or for refreshing it, which is 10000. MAX_KNOW is 20000, which is the maximum turns you can retain a spell. Before there was some oddness in that the spellcasting percents wouldn't reflect your true turns left. Anything over KEEN (10000) would simply be 100%. This does result in players starting with 50%, but I think that's still better than attaining over 100% memory by exercising a spell.