Move castle level so it's right before lethe.
After playtesting with this a bit, I think having the castle where it would normally be in the dungeon is a better layout. The Lethe levels are quite difficult, so it's only fair the player should be able to access the castle first. Getting ahold of the wishes might actually be mandatory in some cases, so let's not box the player out of that.