Fix and refine how Serenity counters spells.
I noticed that the countering was happening a bit deep in the mcastu code, so I brought it out a bit so it can now catch more types of spells. This is a major win for the player. I also penalized the spellcaster a bit more so they get a longer delay when countered.