Towels cannot be worn.
For a while it's been apparent to me that the towel totally outclasses the blindfold. Once one acquires a towel, the blindfold is obsoleted and destined for the junk bin. This commit fixes that.

In practical terms, wearing a towel on your head is not as easy as it would sound. Of course, this depends on the size of the towel, but if we assume the player has a standard hand towel, that would not usually be big enough to fit around a standard head. If we are talking about NetHack players, it's ddebatable if any size towel would be large enough... (Yes, that was a joke, please put the pitchforks away).

In NerfHack, towels have much more utility as removers of grease and blood - making this change a little less severe than it would be in other variants. However, it must be noted that an interesting usage of the wet towel to prevent breathing gas will die with this change. In the interest of propping up the blindfold, I believe this sacrifice is worth the cause.