New filler level before the first level of Sokoban - townfill from UnNetHack.
I really like the idea of this simple buffer level before Soko. I liked it for the Town branch in Un and HackEM, and I think it works equally well here. It adds a new obstacle to overcome before you can get in and it also increases the difficulty of the upper towers.

For my own little touch I added some small pools that are ringed in ice. No grass.