New monster: orb weaver.
This is a new Q class monster, using the transmuter tile from SpliceHack. The orb weaver is capable of summoning elemental spheres to terrorize the player. It also resists every elemental it can cast: fire, cold, shock, poison, and acid.

Although it specializes in summoning orbs, it can also heal itself and cast protection, making it quite defensive and painful to dispose of.

It will need a little playtesting to see where it fits in the dungeon difficulty, but for now I matched it with the genetic engineer.