Allow dungeon features to generate in the Oracle level.
This is pretty close to what I was envisioning for this feature. I felt a bit sad that the Oracle never had any features other than the ones in the middle. It kind of felt like a wasted chance at an altar or other valuable feature, so I enabled them with approximately the same odds as normal dungeon generation. If an altar does spawn here, it will always be cracked though.

I also added an extra room to give a small boost to the chances of something interesting happening.